AccountManagement/requirements.txt
alembic==1.13.1
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
pika==1.3.2
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
